Reorganized about page file structure and working on layout
