+++ frontend/src/components/About/Footer.js
