+++ frontend/src/components/About/TeamCard.js
